MINOR: quic: Add traces about QUIC TLS key update
Dump the secret used to derive the next one during a key update initiated by the
client and dump the resulted new secret and the new key and iv to be used to
decryption Application level packets.

Also add a trace when the key update is supposed to be initiated on haproxy side.

This has already helped in diagnosing an issue evealed by the key update interop
test with xquic as client.

Must be backported to 2.7.

(cherry picked from commit 51a7caf921697d97ff6b5c910797f78fea9cac51)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>