MINOR: quic: consider EBADF as critical on send()
EBADF on sendto() is considered as a fatal error. As such, it is removed
from the list of the transient errors. The connection will be killed
when encountered.

For the record, EBADF can be encountered on process termination with the
listener socket.

This should be backported up to 2.7.

(cherry picked from commit 4bdd0696373b5a683cfab905c65c5af103855bd3)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>