BUG/MEDIUM: fd: make fd_delete() support being called from a different group
There's currently a problem affecting thread groups. Stopping a listener
from a different group than the one that runs this listener will trigger
the BUG_ON() in fd_delete(). This typically happens by issuing "disable
frontend f" on the CLI for the following config since the CLI runs on
group 1:

    global
        nbthread 2
        thread-groups 2
        stats socket /tmp/sock1 level admin

    frontend f
        mode http
        bind abns@frt-sock thread 2

This happens because abns sockets cannot be suspended so here this
requires a full stop.

A first approach would consist in isolating the caller during such rare
operations but it turns out that fd_delete() is not robust against even
such calling conditions, because it uses its own thread mask with an FD
that may be in a different group, and even though the threads would be
isolated and running_mask should be zero, we must not mix thread masks
from different groups like this.

A better solution consists in replacing the bug condition detection with
a self-protection. After all it's not trivial to figure all likely call
places, and forcing upper layers to protect the code is not clean if we
can do it at the bottom. Thus this is what is being done now. We detect
a thread group mismatch, and if so, we forcefully isolate ourselves and
entirely clean the socket. This has the merit of being much more robust
and easier to use (and harder to misuse). Given that such operations are
very rare (actually when they happen a crash follows), it's not a problem
to waste some time isolating the caller there.

This must be backported to 2.7, along with this previous patch:

  BUG/MINOR: fd: used the update list from the fd's group instead of tgid

(cherry picked from commit 061754b249b9903913d6766c1ab31bb393ee5c0d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>