BUG/MINOR: mux-h1: Don't report an error on an early response close
If the response is closed before any data was received, we must not report
an error to the SE descriptor. It is important to be able to retry on an
empty response.

This patch should fix the issue #2061. It must be backported to 2.7.

(cherry picked from commit fb5fff19fee40c660f9839f3bb0b40f33d4899e7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>