BUG/MINOR: http-check: Don't set HTX_SL_F_BODYLESS flag with a log-format body
When the HTTP request of a health-check is forged, we must not pretend there
is no payload, by setting HTX_SL_F_BODYLESS, if a log-format body was
configured.

Indeed, a test on the body length was used but it is only valid for a plain
string. For A log-format string, a list is used. Note it an bug with no
consequence for now.

This patch must be backported as far as 2.2.

(cherry picked from commit 0506d9de512291ed2526654800b98a2317a67b6d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit dd931360c79520aef02394a8027e67cea975b4a2)
Signed-off-by: Willy Tarreau <w@1wt.eu>