BUG/MINOR: http-ana: Don't increment conn_retries counter before the L7 retry
When we are about to perform a L7 retry, we deal with the conn_retries
counter, to be sure we can retry. However, there is an issue here because
the counter is incremented before it is checked against the backend
limit. So, we can miss a connection retry.

Of course, we must invert both operation. The conn_retries counter must be
incremented after the check agains the backend limit.

This patch must be backported as far as 2.6.

(cherry picked from commit 41ade746c727596d5ac6fa10446be25c19d39362)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit f50337c61d0791fcb7ffce2da1ab11cfeac1e692)
Signed-off-by: Willy Tarreau <w@1wt.eu>