MINOR: quic: handle new closing list in show quic
A new global quic-conn list has been added by the previous patch. It will
contain every quic-conn in closing or draining state.

Thus, it is now easier to include or skip them on a "show quic" output :
when the default list on the current thread has been browsed entirely,
either we skip to the next thread or we look at the closing list on the
current thread.

This should be backported up to 2.7.

(cherry picked from commit 2d37629222489d12ced97c1221b78322aba3431a)
Signed-off-by: Willy Tarreau <w@1wt.eu>