BUG/MINOR: init: make sure to always limit the total number of threads
Commit 18521f34fd ("BUG/MINOR: init: properly detect NUMA bindings on
large systems") revealed another interesting issue which is that we can
now fail to start if we have more than 64 threads on a single NUMA node.

2.7 and newer have an explicit test for this and will issue a diag
message indicating this limitation, but 2.6 did not. Let's integrate
the same mechanism so that we always start but print the hard limit.

This will have to be backported to the same versions where the patch
above is backported.