BUG/MINOR: mworker: use MASTER_MAXCONN as default maxconn value
In environments where SYSTEM_MAXCONN is defined when compiling, the
master will use this value instead of the original minimal value which
was set to 100. When this happens, the master process could allocate
RAM excessively since it does not need to have an high maxconn. (For
example if SYSTEM_MAXCONN was set to 100000 or more)

This patch fixes the issue by using the new define MASTER_MAXCONN which
define a default maxconn of 100 for the master process.

Must be backported as far as 2.5.

(cherry picked from commit 2078d4b1f77f70ac110022c2b5ac4644e4c01640)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 585a6cb8ae836094df5ab4944607349866028ca1)
Signed-off-by: Willy Tarreau <w@1wt.eu>